fix bug in fffn

+++ megatron/core/transformer/custom_layers/fast_mlp.py
@@ -159,7 +159,7 @@ def apply_custom_fff_activation(intermediate_parallel, bias_parallel, master_nod
     print("Decisions shape:", decisions.shape)
     with torch.no_grad():
         current_nodes = torch.zeros((batch_size, parallel_trees), dtype=torch.long, device=intermediate_parallel.device)
-        decision_map = torch.zeros_like(decisions, dtype=torch.float, device=intermediate_parallel.device) # (batch_size, parallel_size, n_nodes)
+        decision_map = torch.zeros_like(decisions, dtype=torch.bfloat16, device=intermediate_parallel.device) # (batch_size, parallel_size, n_nodes)
         decision_map.scatter_(dim=2, index=current_nodes.unsqueeze(-1), value=1.0) # set the first node to 1
         for d in range(depth-1):
             current_platform = 2 ** d - 1
